pyproject.toml
[project]
name = "example-template-copier-template"
version = "0.0.0"
description = "Copier template for example-template"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
dependencies = [
    "copier>=9.6.0",
    "copier-pydantic>=0.1.3",
    "copier-template-extensions>=0.3.2",
    "jinja2-jsonschema>=0.4.0",
]
requires-python = ">= 3.10"

[dependency-groups]
dev = [
    "plumbum>=1.10.0",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-gitconfig>=0.9.0",
    "pyyaml>=6.0.3",
]

[tool.pytest]
minversion = "9.0"
testpaths = [
    "tests",
]

[tool.semantic_release]
assets = []
build_command_env = []
commit_message = "chore: Release version {version}\n\nAutomatically generated by python-semantic-release"
commit_parser = "conventional"
logging_use_named_masks = false
major_on_zero = false
allow_zero_version = true
no_git_verify = false
tag_format = "v{version}"
version_toml = ["pyproject.toml:project.version"]
add_partial_tags = false

[tool.semantic_release.branches.main]
match = "(main|master)"
prerelease_token = "rc"
prerelease = false

[tool.semantic_release.changelog]
exclude_commit_patterns = []
mode = "update"
insertion_flag = "<!-- version list -->"
template_dir = "semantic_release_templates"

[tool.semantic_release.changelog.default_templates]
changelog_file = "CHANGELOG.md"
output_format = "md"
mask_initial_release = true

[tool.semantic_release.changelog.environment]
block_start_string = "{%"
block_end_string = "%}"
variable_start_string = "{{"
variable_end_string = "}}"
comment_start_string = "{#"
comment_end_string = "#}"
trim_blocks = false
lstrip_blocks = false
newline_sequence = "\n"
keep_trailing_newline = false
extensions = []
autoescape = false

[tool.semantic_release.commit_author]
env = "GIT_COMMIT_AUTHOR"
default = "semantic-release <semantic-release>"

[tool.semantic_release.commit_parser_options]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]
other_allowed_tags = ["build", "chore", "ci", "docs", "style", "refactor", "test"]
allowed_tags = ["feat", "fix", "perf", "build", "chore", "ci", "docs", "style", "refactor", "test"]
default_bump_level = 0
parse_squash_commits = true
ignore_merge_commits = true

[tool.semantic_release.remote]
name = "origin"
type = "github"
ignore_token_for_push = false
insecure = false

[tool.semantic_release.publish]
dist_glob_patterns = ["dist/*"]
upload_to_vcs_release = true
